feat(mcp): add MCP router starter
